Refactor exception, architecture and delegate store function in new layer
