--- src/main/java/ru/glitchless/auth/validators/IPasswordValidator.java
@@ -1,7 +1,0 @@
-package ru.glitchless.auth.validators;
-
-import ru.glitchless.throwables.InvalidData;
-
-public interface IPasswordValidator {
-    void validate(String rawPassword) throws InvalidData;
-}
