--- src/main/java/ru/glitchless/auth/validators/IUserValidator.java
@@ -1,8 +1,0 @@
-package ru.glitchless.auth.validators;
-
-import ru.glitchless.models.UserModel;
-import ru.glitchless.throwables.InvalidData;
-
-public interface IUserValidator {
-    boolean validate(UserModel user) throws InvalidData;
-}
