+++ src/main/java/ru/glitchless/data/models/UserLocalModel.java
@@ -1,4 +1,4 @@
-package ru.glitchless.models;
+package ru.glitchless.data.models;
 
 public class UserLocalModel {
     private final String login;
