+++ src/main/java/ru/glitchless/data/throwables/NeedAuthorization.java
@@ -1,0 +1,9 @@
+package ru.glitchless.data.throwables;
+
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.UNAUTHORIZED, reason = "For this action you need login.")
+public class NeedAuthorization extends RuntimeException {
+}
