+++ src/main/java/ru/glitchless/data/throwables/UserNotFound.java
@@ -1,0 +1,8 @@
+package ru.glitchless.data.throwables;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "User not found")
+public class UserNotFound extends RuntimeException {
+}
