+++ src/main/java/ru/glitchless/di/ApplicationConfiguration.java
@@ -1,0 +1,22 @@
+package ru.glitchless.di;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import ru.glitchless.utils.MyWebMvcConfigurer;
+
+@Configuration
+public class ApplicationConfiguration {
+
+    @Bean
+    PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    WebMvcConfigurer corsConfigurer() {
+        return new MyWebMvcConfigurer();
+    }
+}
