+++ src/main/java/ru/glitchless/repositories/auth/validators/IPasswordValidator.java
@@ -1,0 +1,5 @@
+package ru.glitchless.repositories.auth.validators;
+
+public interface IPasswordValidator {
+    void validate(String rawPassword);
+}
