+++ src/main/java/ru/glitchless/repositories/auth/validators/IUserValidator.java
@@ -1,0 +1,7 @@
+package ru.glitchless.repositories.auth.validators;
+
+import ru.glitchless.data.models.UserModel;
+
+public interface IUserValidator {
+    void validate(UserModel user);
+}
