--- src/main/java/ru/glitchless/throwables/InvalidData.java
@@ -1,8 +1,0 @@
-package ru.glitchless.throwables;
-
-public class InvalidData extends RuntimeException {
-
-    public InvalidData(String message) {
-        super(message);
-    }
-}
